removed folders to correct package errors
