+++ src/ca/sheridancollege/project/Game.java
@@ -3,7 +3,7 @@
  * Students can modify and extend to implement their game.
  * Add your name as an author and the date!
  */
-package ca.sheridancollege.project.model;
+package ca.sheridancollege.project;
 
 import java.util.ArrayList;
