+++ src/ca/sheridancollege/project/GroupOfCards.java
@@ -3,7 +3,7 @@
  * Students can modify and extend to implement their game.
  * Add your name as an author and the date!
  */
-package ca.sheridancollege.project.model;
+package ca.sheridancollege.project;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -20,7 +20,7 @@ import java.util.Collections;
 public class GroupOfCards {
 
     // The group of cards, stored in an ArrayList
-    private ArrayList<Card> cards;
+    protected ArrayList<Card> cards;
     private int size;// the size of the grouping
 
     public GroupOfCards(int size) {
@@ -33,7 +33,7 @@ public class GroupOfCards {
      *
      * @return the group of cards.
      */
-    public ArrayList<Card> getCards() {
+    protected ArrayList<Card> getCards() {
         return cards;
     }
