+++ src/ca/sheridancollege/project/PokerHand.java
@@ -1,4 +1,4 @@
-package ca.sheridancollege.project.model;
+package ca.sheridancollege.project;
 
 import java.util.ArrayList;
 import java.util.Collections;
