+++ src/ca/sheridancollege/project/PokerPlayer.java
@@ -1,4 +1,4 @@
-package ca.sheridancollege.project.model;
+package ca.sheridancollege.project;
 
 public class PokerPlayer extends Player {
