+++ src/ca/sheridancollege/project/Suit.java
@@ -1,4 +1,4 @@
-package ca.sheridancollege.project.model;
+package ca.sheridancollege.project;
 
 public enum Suit {
     HEARTS, DIAMONDS, CLUBS, SPADES
